# Conflicts:
#	hospital_RabbitMQ_server/src/main/java/com/qf/config/RabbitMailListener.java

+++ hospital_alipay/src/main/java/com/qf/domain/UserOrder.java
@@ -9,7 +9,7 @@ import java.util.Date;
 public class UserOrder {
     private Integer pkOid;
 
-    private Integer userId;
+    private Long userId;
 
     private String docName;
 
@@ -26,7 +26,7 @@ public class UserOrder {
 
     private Date gmtModified;
 
-    public UserOrder(Integer pkOid, Integer userId, String docName, String depName, BigDecimal price, String statement, String orderNum, Date gmtCreate, Date gmtModified) {
+    public UserOrder(Integer pkOid, Long userId, String docName, String depName, BigDecimal price, String statement, String orderNum, Date gmtCreate, Date gmtModified) {
         this.pkOid = pkOid;
         this.userId = userId;
         this.docName = docName;
@@ -50,11 +50,11 @@ public class UserOrder {
         this.pkOid = pkOid;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
